Removed ifs in handle_type_1.c and prepend stuff to print_width(). Changed function ft_putnbr

+++ libft/ft_putnbr.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-void	ft_putnbr(int n)
+void	ft_putnbr(intmax_t n)
 {
 	if (n == -2147483648)
 		ft_putstr("-2147483648");
